Move configs from source code
Configs should live in a separte location from source code.  In this
case we will put it in the root project directory, but will expect them
to live in the ".picontrol" direcotry in the users home.  This is very
standard for many systems like this.

It will also read from the old configs location, but will not write to
it anymore, and will not read the old location if the new location
exists.