Update update feature to use pip
Before it was relying on the distribution file in the git repo to be
updated from time to time, and would actually download the file every
time it even wanted to check for an update.  Now we just as pip if there
is an update, and if there is we install it.  Super simple.