Refactor config.py for use in base project as well
Now config.py is "the one place" for getting information about the
configs.  This means if we need to change things about the way we are
doing comfigs, we only need to modify this file.  No more copying and
pasting into multiple locations.

Also, with statements when manually opening files saves lots of
headaches.

Additionally, some whitespaces cleanup.